+++ konva.js
@@ -8191,8 +8191,8 @@ var Konva = {};
         },
         _useBufferCanvas: function() {
 //            return false;
-            return ((this.getAbsoluteOpacity() !== 1) && this.hasFill() && this.hasStroke()) ||
-               this.hasShadow() && (this.getAbsoluteOpacity() !== 1) && this.hasFill() && this.hasStroke() && this.getStage();
+            return (this.perfectDrawEnabled() && (this.getAbsoluteOpacity() !== 1) && this.hasFill() && this.hasStroke() && this.getStage()) ||
+                   (this.perfectDrawEnabled() && this.hasShadow() && (this.getAbsoluteOpacity() !== 1) && this.hasFill() && this.hasStroke() && this.getStage());
         },
         drawScene: function(can, top) {
             var layer = this.getLayer(),
@@ -8514,6 +8514,25 @@ var Konva = {};
      *
      * // set strokeHitEnabled
      * shape.strokeHitEnabled();
+     */
+
+    Konva.Factory.addGetterSetter(Konva.Shape, 'perfectDrawEnabled', true);
+
+    /**
+     * get/set perfectDrawEnabled. If a shape has fill, stroke and opacity you may set `perfectDrawEnabled` to improve performance.
+     * See http://konvajs.github.io/docs/performance/Disable_Perfect_Draw.html for more information.
+     * Default value is true
+     * @name perfectDrawEnabled
+     * @method
+     * @memberof Konva.Shape.prototype
+     * @param {Boolean} perfectDrawEnabled
+     * @returns {Boolean}
+     * @example
+     * // get perfectDrawEnabled
+     * var perfectDrawEnabled = shape.perfectDrawEnabled();
+     *
+     * // set perfectDrawEnabled
+     * shape.perfectDrawEnabled();
      */
 
     Konva.Factory.addGetterSetter(Konva.Shape, 'lineJoin');
@@ -10336,7 +10355,6 @@ var Konva = {};
          */
         clear: function(bounds) {
             this.getContext().clear(bounds);
-            this.getHitCanvas().getContext().clear(bounds);
             return this;
         },
         clearHitCache: function() {
@@ -10668,26 +10686,8 @@ var Konva = {};
             this.imageData = null; // Clear imageData cache
             return this;
         },
-        /**
-         * clear scene and hit canvas contexts tied to the layer
-         * @method
-         * @memberof Konva.Layer.prototype
-         * @param {Object} [bounds]
-         * @param {Number} [bounds.x]
-         * @param {Number} [bounds.y]
-         * @param {Number} [bounds.width]
-         * @param {Number} [bounds.height]
-         * @example
-         * layer.clear();
-         * layer.clear({
-         *   x : 0,
-         *   y : 0,
-         *   width : 100,
-         *   height : 100
-         * });
-         */
         clear: function(bounds) {
-            this.getContext().clear(bounds);
+            Konva.BaseLayer.prototype.clear.call(this, bounds);
             this.getHitCanvas().getContext().clear(bounds);
             this.imageData = null; // Clear getImageData cache
             return this;
@@ -10803,28 +10803,6 @@ var Konva = {};
         },
         draw: function() {
             this.drawScene();
-            return this;
-        },
-        /**
-         * clear scene and hit canvas contexts tied to the layer
-         * @method
-         * @memberof Konva.FastLayer.prototype
-         * @param {Object} [bounds]
-         * @param {Number} [bounds.x]
-         * @param {Number} [bounds.y]
-         * @param {Number} [bounds.width]
-         * @param {Number} [bounds.height]
-         * @example
-         * layer.clear();
-         * layer.clear({
-         *   x : 0,
-         *   y : 0,
-         *   width : 100,
-         *   height : 100
-         * });
-         */
-        clear: function(bounds) {
-            this.getContext().clear(bounds);
             return this;
         },
         // extend Node.prototype.setVisible